+++ NEWS.md
@@ -1,7 +1,9 @@
 # epiphy 0.4.0
 
-- Fixed sadie algorithm
+- Fixed sadie algorithm.
+- Add mapcomp procedure.
 
 # epiphy 0.3.0
 
-- First release
+- First release.
+-